add VacationApprovers table, change repository, graph

+++ Server/API/Mutations/UserMutation.cs
@@ -18,14 +18,20 @@ public sealed class UserMutation : ObjectGraphType
             .Argument<NonNullGraphType<UserInputGraphType>>("user")
             .ResolveAsync(async context =>
             {
+                var logger = context.RequestServices!.GetRequiredService<ILogger<UserMutation>>();
+                logger.LogInformation("Starting user creation process");
+
                 context.WithPermission(Permission.RegisterUser);
 
                 var user = context.GetArgument<User>("user");
+                logger.LogInformation("Received user data: {@User}", user);
+
                 var userRepository = context.RequestServices!.GetRequiredService<IUserRepository>();
                 var duplicate = await userRepository.GetAsync(user.Username);
 
                 if (duplicate != null)
                 {
+                    logger.LogWarning("User with this username already exists: {Username}", user.Username);
                     throw new ExecutionError("User with this username already exists")
                     {
                         Code = ResponseCode.BadRequest
@@ -33,28 +39,46 @@ public sealed class UserMutation : ObjectGraphType
                 }
 
                 user.PasswordHash = BCrypt.Net.BCrypt.HashPassword(user.PasswordHash);
-                await context.RequestServices!.GetRequiredService<IUserRepository>().InsertAsync(user);
-                await UserMutationExtensions.UpdateApproveVacationsForUsersAsync(user.Username, user.ApprovedVacationsByUsers, userRepository);
+                logger.LogInformation("Password hashed for user: {Username}", user.Username);
+                
+                if (user.VacationApprovers != null)
+                {
+                    user.VacationApprovers.UserUsername = user.Username;
+                    logger.LogInformation("Assigned UserUsername for VacationApprovers: {Username}", user.Username);
+                }
+
+                await userRepository.InsertAsync(user);
+                logger.LogInformation("User successfully created: {Username}", user.Username);
+
                 return user;
             });
 
-
+        
         Field<UserGraphType>("delete")
             .Argument<NonNullGraphType<StringGraphType>>("username")
             .ResolveAsync(async context =>
             {
+                var logger = context.RequestServices!.GetRequiredService<ILogger<UserMutation>>();
+                logger.LogInformation("Starting user deletion process");
+
                 context.WithPermission(Permission.DeleteUser);
 
                 var username = context.GetArgument<string>("username");
                 var repository = context.RequestServices!.GetRequiredService<IUserRepository>();
-                var user = await repository.GetAsync(username)
-                    ?? throw new ExecutionError("User not found") { Code = ResponseCode.BadRequest };
+                var user = await repository.GetAsync(username);
+
+                if (user == null)
+                {
+                    logger.LogWarning("User not found: {Username}", username);
+                    throw new ExecutionError("User not found") { Code = ResponseCode.BadRequest };
+                }
 
                 await repository.DeleteAsync(user.Username);
-                await UserMutationExtensions.RemoveApproveVacationsForUsersAsync(user.Username, repository);
+                logger.LogInformation("User successfully deleted: {Username}", username);
+
                 return user;
             });
-
+        
         Field<UserGraphType>("update")
             .Argument<NonNullGraphType<UserUpdateInputGraphType>>("user")
             .ResolveAsync(async context =>
@@ -64,16 +88,21 @@ public sealed class UserMutation : ObjectGraphType
                 var requestUser = context.GetArgument<User>("user");
                 var repository = context.RequestServices!.GetRequiredService<IUserRepository>();
                 var oldUser = await repository.GetAsync(requestUser.Username)
-                    ?? throw new ExecutionError("User not found") { Code = ResponseCode.BadRequest };
+                              ?? throw new ExecutionError("User not found") { Code = ResponseCode.BadRequest };
 
                 var username = context.User?.FindFirstValue(ClaimTypes.NameIdentifier);
 
                 if (username == oldUser.Username)
                 {
-                    throw new ExecutionError("Cannot modify own yourself")
+                    throw new ExecutionError("Cannot modify yourself")
                     {
                         Code = ResponseCode.BadRequest
                     };
+                }
+                
+                if (oldUser.VacationApprovers != null)
+                {
+                    oldUser.VacationApprovers.UserUsername = requestUser.Username;
                 }
 
                 var newUser = new User
@@ -88,14 +117,14 @@ public sealed class UserMutation : ObjectGraphType
                     State = oldUser.State,
                     WorkType = requestUser.WorkType,
                     WorkTime = requestUser.WorkTime,
-                    ApprovedVacationsByUsers = requestUser.ApprovedVacationsByUsers
+                    VacationApprovers = new VacationApprovers
+                    {
+                        UserUsername = requestUser.Username,
+                        ApprovedVacationsByUsers = requestUser.VacationApprovers?.ApprovedVacationsByUsers ?? []
+                    }
                 };
 
                 await repository.UpdateAsync(newUser);
-                await UserMutationExtensions.UpdateApproveVacationsForUsersAsync(
-                    newUser.Username, 
-                    newUser.ApprovedVacationsByUsers, 
-                    repository);
                 return newUser;
             });
